+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased changes
 
 - Add `NextUpdateSequenceNumbers::protocol_level_tokens` and protobuf deserialization of it
+- Changed `TokenClient`'s `burn` and `mint` methods to accept a singular `TokenAmount`, instead of `Vec<TokenAmount>`.
 
 ## 7.0.0-alpha.3